Move upload flow logic to a service
